Move incentives fetching to useIncentives
